priority delete: if there are related tickets, migrate them to another existing one

+++ src/Controllers/PrioritiesController.php
@@ -5,6 +5,7 @@ namespace PanicHD\PanicHD\Controllers;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
+use PanicHD\PanicHD\Helpers\LaravelVersion;
 use PanicHD\PanicHD\Models\Priority;
 
 class PrioritiesController extends Controller
@@ -16,11 +17,15 @@ class PrioritiesController extends Controller
      */
     public function index()
     {
-        $priorities = \Cache::remember('panichd::priorities', 60, function () {
-            return Priority::all();
-        });
+        $priorities = Priority::withCount('tickets')->get();
+		
+		if (LaravelVersion::min('5.3.0')) {
+            $priorities_list = $priorities->pluck('name', 'id')->toArray();
+        } else {
+            $priorities_list = $priorities->lists('name', 'id')->toArray();
+        }
 
-        return view('panichd::admin.priority.index', compact('priorities'));
+        return view('panichd::admin.priority.index', compact('priorities', 'priorities_list'));
     }
 
     /**
@@ -51,9 +56,6 @@ class PrioritiesController extends Controller
         $priority->create(['name' => $request->name, 'color' => $request->color]);
 
         Session::flash('status', trans('panichd::lang.priority-name-has-been-created', ['name' => $request->name]));
-
-        \Cache::forget('panichd::priorities');
-
         return redirect()->action('\PanicHD\PanicHD\Controllers\PrioritiesController@index');
     }
 
@@ -102,9 +104,6 @@ class PrioritiesController extends Controller
         $priority->update(['name' => $request->name, 'color' => $request->color]);
 
         Session::flash('status', trans('panichd::lang.priority-name-has-been-modified', ['name' => $request->name]));
-
-        \Cache::forget('panichd::priorities');
-
         return redirect()->action('\PanicHD\PanicHD\Controllers\PrioritiesController@index');
     }
 
@@ -115,16 +114,25 @@ class PrioritiesController extends Controller
      *
      * @return Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
         $priority = Priority::findOrFail($id);
         $name = $priority->name;
-        $priority->delete();
+		
+		if ($request->has('tickets_new_priority_id')){
+			$this->validate($request, [
+				'tickets_new_priority_id' => 'required|exists:panichd_priorities,id',
+			]);
+			$priority->delete($request->tickets_new_priority_id);
+		}else{
+			if ($priority->tickets()->count() > 0){
+				return back()->with('warning', trans('panichd::admin.priority-delete-error-no-priority', ['name' => $name]));
+			}
+			
+			$priority->delete();
+		}
 
         Session::flash('status', trans('panichd::lang.priority-name-has-been-deleted', ['name' => $name]));
-
-        \Cache::forget('panichd::priorities');
-
         return redirect()->action('\PanicHD\PanicHD\Controllers\PrioritiesController@index');
     }
 }
